% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabt.R
\docType{class}
\name{RangedBT-class}
\alias{RangedBT-class}
\title{A class for manipulating BigTable-resident assay and range metadata on features.}
\description{
A class for manipulating BigTable-resident assay and range metadata on features.
}
\section{Slots}{

\describe{
\item{\code{assay}}{a \code{tbl_bigquery} instance/reference for a rectangular table}

\item{\code{rangeData}}{a \code{tbl_bigquery} instance/reference for a rectangular table with information about rows of \code{assay}, including all information required to define a \code{\link[GenomicRanges]{GRanges-class}} instance}

\item{\code{project}}{a character string naming the Google Cloud Platform project in use}

\item{\code{cached_nrow}}{Because counting rows can be expensive, this integer slot is available for setting by \code{\link{set_cached_nrow}}}
}}

