% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabt.R
\name{RangedBT}
\alias{RangedBT}
\title{Simple constructor for RangedBT}
\usage{
RangedBT(project, src, assayname, rangename)
}
\arguments{
\item{project}{character tag for Google Cloud Platform project}

\item{src}{'src_bigquery' instance}

\item{assayname}{name of table with assay data}

\item{rangename}{name of table with range meta data}
}
\description{
Simple constructor for RangedBT
}

