% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabt.R
\docType{methods}
\name{RangedBTmethods}
\alias{RangedBTmethods}
\alias{assay,RangedBT,ANY-method}
\alias{assayname}
\alias{datasetname}
\alias{nrow,RangedBT-method}
\alias{project}
\alias{project,RangedBT-method}
\alias{rowRanges,RangedBT-method}
\alias{set_cached_nrow}
\alias{show,RangedBT-method}
\title{Methods for RangedBT}
\usage{
\S4method{assay}{RangedBT,ANY}(x, i, ...)

\S4method{rowRanges}{RangedBT}(x, ...)

project(x)

\S4method{project}{RangedBT}(x)

assayname(x)

datasetname(x)

\S4method{show}{RangedBT}(object)

\S4method{nrow}{RangedBT}(x)

set_cached_nrow(x, nr)
}
\arguments{
\item{x}{RangedBT instance}

\item{i}{ignored}

\item{object}{a RangedBT instance}

\item{nr}{integer row count, not checked}

\item{\dots}{ignored}
}
\description{
Basic accessors

simple accessors for assay and metadata components

extract character name of dataset in which RangedBT is defined
}

